django_lifeplan_genius/requirements.txt
# Django and REST framework
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0

# Database
psycopg2-binary==2.9.10

# Environment variables
python-dotenv==1.0.1

# Document generation
python-docx==1.1.2
openpyxl==3.1.5

# Utilities
python-dateutil==2.9.0
